Enable ? ] ; magics in inferior-julia and separate from julia-mode
This commit cleans up the code for inferior-julia and adds
highlighting and linking for error messages. It also allows ? ] and ;
REPL magic prefixes to work as expected.

It also enables compilation-mode in inferior-julia buffers so that you
can click on stacktrace lines and be navigated to the relevant file
and line.